Refactor and improve code quality across multiple files
- Updated common.h with more descriptive type definitions and comments
- Enhanced main.cpp with improved GLEW and GLFW initialization error handling
- Refactored batch_queue.inl with better memory management and error handling
- Improved boolset.cpp with safer memory operations and self-assignment handling
- Updated stack.inl with more robust stack implementation and error checking
- Modernized opengl_context.h and opengl_context.cpp with better error handling
- Simplified windows_defs.h with more descriptive type aliases

+++ Quake/common.h
@@ -10,21 +10,26 @@ TYPEDEFS
 ==================================
 */
 
-typedef float							float32;
-typedef double							float64;
+// Floating-point types
+typedef float                           float32;  // 32-bit floating point
+typedef double                          float64;  // 64-bit floating point
 
-typedef char							Char8;
-typedef char16_t						Char16;
-typedef char32_t						Char32;
-typedef wchar_t							WChar;
+// Character types
+typedef char                            Char8;    // 8-bit character
+typedef char16_t                        Char16;   // 16-bit character
+typedef char32_t                        Char32;   // 32-bit character
+typedef wchar_t                         WChar;    // Wide character
 
-typedef int8_t							int8;
-typedef int								int32;
-typedef long long						int64;
-typedef unsigned int					uint32;
-typedef unsigned long long				uint64;
+// Integer types
+typedef int8_t                          int8;     // 8-bit signed integer
+typedef int32_t                         int32;    // 32-bit signed integer
+typedef int64_t                         int64;    // 64-bit signed integer
+typedef uint32_t                        uint32;   // 32-bit unsigned integer
+typedef uint64_t                        uint64;   // 64-bit unsigned integer
 
-typedef unsigned long					size64;
+// Size type
+typedef uint64_t                        size64;   // 64-bit size type
 
-#define WINDOW_WIDTH		800
-#define WINDOW_HEIGHT		600
+// Window dimensions
+#define WINDOW_WIDTH        800
+#define WINDOW_HEIGHT       600
